move all types into types.rs
