+++ proofs/fstar/extraction/Libcrux.Kem.Kyber.Constant_time_ops.fst
@@ -6,7 +6,7 @@ let is_non_zero (value: u8) : u8 =
   let value_negated:i8 = Core.Ops.Arith.Neg.neg (cast value <: i8) in
   ((value |. (cast value_negated <: u8) <: u8) >>! 7l <: u8) &. 1uy
 
-let compare_ciphertexts_in_constant_time (#v_CIPHERTEXT_SIZE: usize) (lhs rhs: slice u8) : u8 =
+let compare_ciphertexts_in_constant_time (#v_CIPHERTEXT_SIZE: usize) (lhs rhs: t_Slice u8) : u8 =
   let _:Prims.unit =
     if true
     then
@@ -45,18 +45,18 @@ let compare_ciphertexts_in_constant_time (#v_CIPHERTEXT_SIZE: usize) (lhs rhs: s
   in
   let (r: u8):u8 = 0uy in
   let r:u8 =
-    Core.Iter.Traits.Iterator.Iterator.fold (Core.Iter.Traits.Collect.f_into_iter ({
+    Core.Iter.Traits.Iterator.f_fold (Core.Iter.Traits.Collect.f_into_iter ({
               Core.Ops.Range.f_start = sz 0;
               Core.Ops.Range.f_end = v_CIPHERTEXT_SIZE
             })
         <:
-        (Core.Iter.Traits.Collect.impl (Core.Ops.Range.t_Range usize)).f_IntoIter)
+        Core.Ops.Range.t_Range usize)
       r
       (fun r i -> r |. ((lhs.[ i ] <: u8) ^. (rhs.[ i ] <: u8) <: u8) <: u8)
   in
   is_non_zero r
 
-let select_shared_secret_in_constant_time (lhs rhs: slice u8) (selector: u8) : array u8 (sz 32) =
+let select_shared_secret_in_constant_time (lhs rhs: t_Slice u8) (selector: u8) : t_Array u8 (sz 32) =
   let _:Prims.unit =
     if true
     then
@@ -93,15 +93,15 @@ let select_shared_secret_in_constant_time (lhs rhs: slice u8) (selector: u8) : a
       in
       ()
   in
-  let mask:u8 = Core.Num.impl_6__wrapping_sub (is_non_zero selector <: u8) 1uy in
-  let out:array u8 (sz 32) = Rust_primitives.Hax.repeat 0uy (sz 32) in
-  let out:array u8 (sz 32) =
-    Core.Iter.Traits.Iterator.Iterator.fold (Core.Iter.Traits.Collect.f_into_iter ({
+  let mask:u8 = Core.Num.impl__u8__wrapping_sub (is_non_zero selector <: u8) 1uy in
+  let out:t_Array u8 (sz 32) = Rust_primitives.Hax.repeat 0uy (sz 32) in
+  let out:t_Array u8 (sz 32) =
+    Core.Iter.Traits.Iterator.f_fold (Core.Iter.Traits.Collect.f_into_iter ({
               Core.Ops.Range.f_start = sz 0;
               Core.Ops.Range.f_end = Libcrux.Kem.Kyber.Constants.v_SHARED_SECRET_SIZE
             })
         <:
-        (Core.Iter.Traits.Collect.impl (Core.Ops.Range.t_Range usize)).f_IntoIter)
+        Core.Ops.Range.t_Range usize)
       out
       (fun out i ->
           Rust_primitives.Hax.update_at out
@@ -113,6 +113,6 @@ let select_shared_secret_in_constant_time (lhs rhs: slice u8) (selector: u8) : a
               <:
               u8)
           <:
-          array u8 (sz 32))
+          t_Array u8 (sz 32))
   in
   out
