+++ proofs/fstar/extraction/Libcrux.Kem.Kyber.Conversions.fst
@@ -2,7 +2,7 @@ module Libcrux.Kem.Kyber.Conversions
 #set-options "--fuel 0 --ifuel 1 --z3rlimit 15"
 open Core
 
-let into_padded_array (#v_LEN: usize) (slice: slice u8) : array u8 v_LEN =
+let into_padded_array (#v_LEN: usize) (slice: t_Slice u8) : t_Array u8 v_LEN =
   let _:Prims.unit =
     if true
     then
@@ -16,8 +16,8 @@ let into_padded_array (#v_LEN: usize) (slice: slice u8) : array u8 v_LEN =
       in
       ()
   in
-  let out:array u8 v_LEN = Rust_primitives.Hax.repeat 0uy v_LEN in
-  let out:array u8 v_LEN =
+  let out:t_Array u8 v_LEN = Rust_primitives.Hax.repeat 0uy v_LEN in
+  let out:t_Array u8 v_LEN =
     Rust_primitives.Hax.update_at out
       ({ Core.Ops.Range.f_start = sz 0; Core.Ops.Range.f_end = Core.Slice.impl__len slice <: usize }
       )
@@ -27,33 +27,33 @@ let into_padded_array (#v_LEN: usize) (slice: slice u8) : array u8 v_LEN =
                   Core.Ops.Range.f_end = Core.Slice.impl__len slice <: usize
                 })
             <:
-            slice u8)
+            t_Slice u8)
           slice
         <:
-        slice u8)
+        t_Slice u8)
   in
   out
 
 class t_UpdatingArray (#v_Self: Type) = {
   [@@@ FStar.Tactics.Typeclasses.no_method]_super_509883233:t_UpdatingArray v_Self;
-  f_push:v_Self -> slice u8 -> v_Self
+  f_push:v_Self -> t_Slice u8 -> v_Self
 }
 
 type t_UpdatableArray (#v_LEN: usize) = {
-  f_value:array u8 v_LEN;
+  f_value:t_Array u8 v_LEN;
   f_pointer:usize
 }
 
-let impl__new (#v_LEN: usize) (value: array u8 v_LEN) : t_UpdatableArray v_LEN =
+let impl__new (#v_LEN: usize) (value: t_Array u8 v_LEN) : t_UpdatableArray v_LEN =
   { f_value = value; f_pointer = sz 0 }
 
-let impl__array (#v_LEN: usize) (self: t_UpdatableArray v_LEN) : array u8 v_LEN = self.f_value
+let impl__array (#v_LEN: usize) (self: t_UpdatableArray v_LEN) : t_Array u8 v_LEN = self.f_value
 
 let impl_1 (#v_LEN: usize) : t_UpdatingArray (t_UpdatableArray v_LEN) =
   {
-    f_impl_1__push
+    f_push
     =
-    fun (#v_LEN: usize) (self: t_UpdatableArray v_LEN) (other: slice u8) ->
+    fun (#v_LEN: usize) (self: t_UpdatableArray v_LEN) (other: t_Slice u8) ->
       let self:t_UpdatableArray v_LEN =
         {
           self with
@@ -74,10 +74,10 @@ let impl_1 (#v_LEN: usize) : t_UpdatingArray (t_UpdatableArray v_LEN) =
                         self.f_pointer +! (Core.Slice.impl__len other <: usize) <: usize
                       })
                   <:
-                  slice u8)
+                  t_Slice u8)
                 other
               <:
-              slice u8)
+              t_Slice u8)
         }
       in
       let self:t_UpdatableArray v_LEN =
