+++ src/kem/kyber/kyber768.rs
@@ -39,10 +39,7 @@ pub type Kyber768SharedSecret = KyberSharedSecret<SHARED_SECRET_SIZE>;
 /// Generate Kyber 768 Key Pair
 pub fn generate_key_pair_768(
     randomness: [u8; KEY_GENERATION_SEED_SIZE],
-) -> Result<
-    KyberKeyPair<SECRET_KEY_SIZE_768, CPA_PKE_PUBLIC_KEY_SIZE_768>,
-    BadRejectionSamplingRandomnessError,
-> {
+) -> Result<KyberKeyPair<SECRET_KEY_SIZE_768, CPA_PKE_PUBLIC_KEY_SIZE_768>, Error> {
     generate_keypair::<
         RANK_768,
         CPA_PKE_SECRET_KEY_SIZE_768,
@@ -63,7 +60,7 @@ pub fn encapsulate_768(
         KyberCiphertext<CPA_PKE_CIPHERTEXT_SIZE_768>,
         KyberSharedSecret<SHARED_SECRET_SIZE>,
     ),
-    BadRejectionSamplingRandomnessError,
+    Error,
 > {
     encapsulate::<
         RANK_768,
