+++ src/kem/kyber/types.rs
@@ -1,5 +1,3 @@
-use super::{KyberKeyPair, KyberPrivateKey, KyberPublicKey};
-
 macro_rules! impl_generic_struct {
     ($name:ident) => {
         pub struct $name<const SIZE: usize> {
@@ -80,6 +78,20 @@ macro_rules! impl_generic_struct {
             }
         }
     };
+}
+
+impl_generic_struct!(KyberCiphertext);
+impl_generic_struct!(KyberSharedSecret);
+impl_generic_struct!(KyberPrivateKey);
+impl_generic_struct!(KyberPublicKey);
+
+// The PKE Private Key
+impl_generic_struct!(PrivateKey);
+
+/// A Kyber key pair
+pub struct KyberKeyPair<const PRIVATE_KEY_SIZE: usize, const PUBLIC_KEY_SIZE: usize> {
+    pub(crate) sk: KyberPrivateKey<PRIVATE_KEY_SIZE>,
+    pub(crate) pk: KyberPublicKey<PUBLIC_KEY_SIZE>,
 }
 
 impl<const PRIVATE_KEY_SIZE: usize, const PUBLIC_KEY_SIZE: usize>
